rename media queries file
